Safari-609.1.11.1/LayoutTests/accessibility/mac/webkit-scrollarea-expected.txt

text

text
This tests that the WebKit generated scroll areas are correct.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS body.role is 'AXRole: AXWebArea'
PASS body.parentElement().role is 'AXRole: AXScrollArea'
PASS body.parentElement().childrenCount is 3
PASS successfullyParsed is true

TEST COMPLETE
